+++ www/facebook-browser.js
@@ -24,10 +24,7 @@ exports.showDialog = function showDialog (options, s, f) {
 
   options.name = options.name || ''
   options.message = options.message || ''
-  options.caption = options.caption || ''
-  options.description = options.description || ''
   options.href = options.href || ''
-  options.picture = options.picture || ''
   options.quote = options.quote || ''
   if (options.actionType) {
     options.action_type = options.actionType
@@ -134,7 +131,7 @@ exports.getAccessToken = function getAccessToken (s, f) {
     if(s) s(response);
     return
   }
-  if(f) f('NO_TOKEN');
+  if(f) f('Session not open.');
 }
 
 exports.logEvent = function logEvent (eventName, params, valueToSum, s, f) {
@@ -205,11 +202,6 @@ exports.api = function api (graphPath, permissions, httpMethod, s, f) {
       if(s) s(response);
     }
   })
-}
-
-exports.browserInit = function browserInit (appId, version, s) {
-  console.warn("browserInit is deprecated and may be removed in the future");
-  console.trace();
 }
 
 exports.activateApp = function logEvent (s, f) {